Bug fixes in fitPHospModel() and processData()
In fitPHospModel(), the simplified binomial model is only used if there are zeros in the pHosp data *during the relevant fitting window* as opposed to anywhere in the entire dataset.

In processData(),  if epi data is more recent that hospital occupancy data (defined by latest date present in table) then hospital occupancy data is padded with NaNs instead of discarding the additional epi data.